MAPSJS-2660: Move Object3D size estimation to internal namespace.
