+++ @here/harp-mapview/lib/Utils.ts
@@ -18,24 +18,15 @@ import {
 } from "@here/harp-geoutils";
 import { GeoCoordLike } from "@here/harp-geoutils/lib/coordinates/GeoCoordLike";
 import { EarthConstants } from "@here/harp-geoutils/lib/projection/EarthConstants";
-import { MapMeshBasicMaterial, MapMeshStandardMaterial } from "@here/harp-materials";
 import { assert, LoggerManager } from "@here/harp-utils";
 import * as THREE from "three";
 
 import { CameraUtils } from "./CameraUtils";
 import { ElevationProvider } from "./ElevationProvider";
-import { LodMesh } from "./geometry/LodMesh";
+import { Object3DUtils } from "./geometry/Object3DUtils";
 import { MapView } from "./MapView";
-import { getFeatureDataSize, TileFeatureData } from "./Tile";
 
 const logger = LoggerManager.instance.create("MapViewUtils");
-
-// Estimation of the size of an Object3D with all the simple properties, like matrices and flags.
-// There may be cases where it is possible to construct Object3Ds with considerable less memory
-// consumption, but this value is used to simplify the estimation.
-const MINIMUM_OBJECT3D_SIZE_ESTIMATION = 1000;
-
-const MINIMUM_ATTRIBUTE_SIZE_ESTIMATION = 56;
 
 /**
  * Zoom level to request terrain tiles for getting the height of the camera above terrain.
@@ -131,12 +122,9 @@ export namespace MapViewUtils {
     }
 
     /**
-     * Describes estimated usage of memory on heap and GPU.
+     * @deprecated
      */
-    export interface MemoryUsage {
-        heapSize: number;
-        gpuSize: number;
-    }
+    export interface MemoryUsage extends Object3DUtils.MemoryUsage {}
 
     /**
      * Zooms and moves the map in such a way that the given target position remains at the same
@@ -1732,41 +1720,9 @@ export namespace MapViewUtils {
     export const calculateWorldSizeByFocalLength = CameraUtils.convertScreenToWorldSize;
 
     /**
-     * Computes estimate for size of a THREE.Object3D object and its children. Shared materials
-     * and/or attributes will be counted multiple times.
-     *
-     * @param object - The mesh object to evaluate
-     * @param size - The {@link MemoryUsage} to update.
-     * @param visitedObjects - Optional map to store large objects that could be shared.
-     *
-     * @returns Estimate of object size in bytes for heap and GPU.
+     * @deprecated
      */
-    export function estimateObject3dSize(
-        object: THREE.Object3D,
-        parentSize?: MemoryUsage,
-        visitedObjects?: Map<string, boolean>
-    ): MemoryUsage {
-        const size =
-            parentSize !== undefined
-                ? parentSize
-                : {
-                      heapSize: 0,
-                      gpuSize: 0
-                  };
-
-        if (visitedObjects === undefined) {
-            visitedObjects = new Map();
-        }
-
-        estimateMeshSize(object, size, visitedObjects);
-
-        if (object.children.length > 0) {
-            for (const child of object.children) {
-                estimateObject3dSize(child, size, visitedObjects);
-            }
-        }
-        return size;
-    }
+    export const estimateObject3dSize = Object3DUtils.estimateSize;
 
     /**
      * Check if tiles or other content is currently being loaded.
@@ -1815,221 +1771,6 @@ export namespace MapViewUtils {
             return true;
         }
         return false;
-    }
-
-    function estimateTextureSize(
-        texture: THREE.Texture | null,
-        objectSize: MemoryUsage,
-        visitedObjects: Map<string, boolean>
-    ): void {
-        if (
-            texture === null ||
-            texture === undefined ||
-            texture.image === undefined ||
-            texture.image === null
-        ) {
-            return;
-        }
-
-        if (texture.uuid !== undefined && visitedObjects.get(texture.uuid) === true) {
-            return;
-        }
-        visitedObjects.set(texture.uuid, true);
-
-        // May be HTMLImage or ImageData
-        const image = texture.image;
-        // Assuming RGBA
-        const imageBytes = 4 * image.width * image.height;
-        objectSize.heapSize += imageBytes;
-        objectSize.gpuSize += imageBytes;
-    }
-
-    function estimateMaterialSize(
-        material: THREE.Material,
-        objectSize: MemoryUsage,
-        visitedObjects: Map<string, boolean>
-    ): void {
-        if (material.uuid !== undefined && visitedObjects.get(material.uuid) === true) {
-            return;
-        }
-        visitedObjects.set(material.uuid, true);
-
-        if (
-            material instanceof THREE.RawShaderMaterial ||
-            material instanceof THREE.ShaderMaterial
-        ) {
-            const rawMaterial = material;
-            for (const name in rawMaterial.uniforms) {
-                if (rawMaterial.uniforms[name] !== undefined) {
-                    const uniform = rawMaterial.uniforms[name];
-                    if (uniform instanceof THREE.Texture) {
-                        estimateTextureSize(uniform, objectSize, visitedObjects);
-                    }
-                }
-            }
-        } else if (
-            material instanceof THREE.MeshBasicMaterial ||
-            material instanceof MapMeshBasicMaterial
-        ) {
-            const meshMaterial = material;
-            estimateTextureSize(meshMaterial.map, objectSize, visitedObjects);
-            estimateTextureSize(meshMaterial.aoMap, objectSize, visitedObjects);
-            estimateTextureSize(meshMaterial.specularMap, objectSize, visitedObjects);
-            estimateTextureSize(meshMaterial.alphaMap, objectSize, visitedObjects);
-            estimateTextureSize(meshMaterial.envMap, objectSize, visitedObjects);
-        } else if (material instanceof MapMeshStandardMaterial) {
-            const standardMaterial = material;
-
-            estimateTextureSize(standardMaterial.map, objectSize, visitedObjects);
-            estimateTextureSize(standardMaterial.lightMap, objectSize, visitedObjects);
-            estimateTextureSize(standardMaterial.aoMap, objectSize, visitedObjects);
-            estimateTextureSize(standardMaterial.emissiveMap, objectSize, visitedObjects);
-            estimateTextureSize(standardMaterial.bumpMap, objectSize, visitedObjects);
-            estimateTextureSize(standardMaterial.normalMap, objectSize, visitedObjects);
-            estimateTextureSize(standardMaterial.displacementMap, objectSize, visitedObjects);
-            estimateTextureSize(standardMaterial.roughnessMap, objectSize, visitedObjects);
-            estimateTextureSize(standardMaterial.metalnessMap, objectSize, visitedObjects);
-            estimateTextureSize(standardMaterial.alphaMap, objectSize, visitedObjects);
-            estimateTextureSize(standardMaterial.envMap, objectSize, visitedObjects);
-        } else if (
-            material instanceof THREE.LineBasicMaterial ||
-            material instanceof THREE.LineDashedMaterial ||
-            material instanceof THREE.PointsMaterial
-        ) {
-            // Nothing to be done here
-        } else {
-            logger.warn("estimateMeshSize: unidentified material: ", material);
-        }
-    }
-
-    function estimateAttributeSize(
-        attribute: any,
-        attrName: string,
-        objectSize: MemoryUsage,
-        visitedObjects: Map<string, boolean>
-    ): void {
-        // Attributes (apparently) do not have their uuid set up.
-        if (attribute.uuid === undefined) {
-            attribute.uuid = THREE.MathUtils.generateUUID();
-        }
-
-        if (visitedObjects.get(attribute.uuid) === true) {
-            return;
-        }
-        visitedObjects.set(attribute.uuid, true);
-
-        let attrBytes = 0;
-        let bytesPerElement = 4;
-        if (attribute.array.BYTES_PER_ELEMENT !== undefined) {
-            bytesPerElement = attribute.array.BYTES_PER_ELEMENT;
-        }
-        if (
-            attribute instanceof THREE.InterleavedBufferAttribute ||
-            attribute instanceof THREE.BufferAttribute
-        ) {
-            attrBytes = bytesPerElement * attribute.count * attribute.itemSize;
-        } else {
-            logger.warn("estimateMeshSize: unidentified attribute: ", attrName);
-        }
-
-        objectSize.heapSize += attrBytes + MINIMUM_ATTRIBUTE_SIZE_ESTIMATION;
-        objectSize.gpuSize += attrBytes;
-    }
-
-    function estimateGeometrySize(
-        geometry: THREE.BufferGeometry,
-        objectSize: MemoryUsage,
-        visitedObjects: Map<string, boolean>
-    ): void {
-        const isNewObject =
-            geometry.uuid === undefined || visitedObjects.get(geometry.uuid) !== true;
-
-        if (!isNewObject) {
-            return;
-        }
-        visitedObjects.set(geometry.uuid, true);
-
-        if (geometry === undefined) {
-            // Nothing more to calculate.
-            return;
-        }
-
-        const attributes = geometry.attributes;
-        if (attributes === undefined) {
-            logger.warn("estimateGeometrySize: unidentified geometry: ", geometry);
-            return;
-        }
-
-        for (const property in attributes) {
-            if (attributes[property] !== undefined) {
-                estimateAttributeSize(attributes[property], property, objectSize, visitedObjects);
-            }
-        }
-        if (geometry.index !== null) {
-            estimateAttributeSize(geometry.index, "index", objectSize, visitedObjects);
-        }
-    }
-
-    function estimateMeshSize(
-        object: THREE.Object3D,
-        objectSize: MemoryUsage,
-        visitedObjects: Map<string, boolean>
-    ): void {
-        if (!object.isObject3D || object instanceof THREE.Scene) {
-            return;
-        }
-
-        if (object.uuid !== undefined && visitedObjects.get(object.uuid) === true) {
-            return;
-        }
-        visitedObjects.set(object.uuid, true);
-
-        if ((object as any).isMesh || (object as any).isLine || (object as any).isPoints) {
-            // Estimated minimum impact on heap.
-            let heapSize = MINIMUM_OBJECT3D_SIZE_ESTIMATION;
-            const gpuSize = 0;
-
-            // Cast to LodMesh class which contains the minimal required properties sub-set.
-            const mesh = object as LodMesh;
-
-            // Calculate material(s) impact.
-            if (mesh.material !== undefined) {
-                if (Array.isArray(mesh.material)) {
-                    const materials = mesh.material as THREE.Material[];
-                    for (const material of materials) {
-                        estimateMaterialSize(material, objectSize, visitedObjects);
-                    }
-                } else {
-                    const material = mesh.material as THREE.Material;
-                    estimateMaterialSize(material, objectSize, visitedObjects);
-                }
-            }
-
-            // Calculate cost of geometry.
-            if (mesh.geometries !== undefined) {
-                for (const geometry of mesh.geometries) {
-                    estimateGeometrySize(geometry, objectSize, visitedObjects);
-                }
-            } else if (mesh.geometry !== undefined) {
-                estimateGeometrySize(mesh.geometry, objectSize, visitedObjects);
-            }
-
-            // Add info that is required for picking (parts of) objects and match them to
-            // the featureID in the map data.
-            const featureData: TileFeatureData | undefined =
-                object.userData !== undefined
-                    ? (object.userData.feature as TileFeatureData)
-                    : undefined;
-
-            if (featureData !== undefined) {
-                heapSize += getFeatureDataSize(featureData);
-            }
-
-            objectSize.heapSize += heapSize;
-            objectSize.gpuSize += gpuSize;
-        } else {
-            logger.warn("estimateMeshSize: unidentified object", object);
-        }
     }
 
     /**
